+++ src/app/pipes/duration.pipe.ts
