+++ src/app/pipes/search.pipe.ts
